RAG Q%A Based System/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-genai
PyPDF2

